feat(#384): merge StaleAdminsPanel and MemberPermissionDistributionPanel into combined panel
Absorbs member permission distribution UI into StaleAdminsPanel, rendering
permission summary (admins/members/collaborators) at top and admin activity
detail as a collapsible sub-section below. Reduces Governance tab from three
panels to two, eliminating the redundant admin count display and the parent-level
useStaleAdmins plumbing in GovernanceExtraPanels.
